Update Swedish Logos (#620)
* Change logos

* Fixes
