+++ src/core/Config.ts
@@ -5,7 +5,6 @@ import DefaultConfig from './config/default.json';
 
 export type ConfigType = {
   country: string;
-  showPartnerLogos: boolean;
   enableMultiplePatients: boolean;
   enablePersonalInformation: boolean;
   enableCohorts: boolean;
