+++ src/core/config/GB.json
@@ -1,6 +1,4 @@
 {
     "country": "GB",
-    "showPartnerLogos": false,
-
     "defaultHeightUnit": "ft"
 }
