+++ src/core/config/default.json
@@ -1,6 +1,5 @@
 {
     "country": "ZZ",
-    "showPartnerLogos": true,
     "enableMultiplePatients": true,
     "enablePersonalInformation": true,
     "enableCohorts": true,
