+++ src/storybook/stories/PoweredByZoe.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react-native';
 
-import { PoweredByZoe } from '@covid/components/PoweredByZoe';
+import { PoweredByZoe } from '@covid/components/Logos/PoweredByZoe';
 import { CenterView, DarkBackground } from '@covid/storybook/decorator';
 
 storiesOf('PoweredByZoe', module)
